More casts to unsigned int.

+++ 1Assignment/src/blur_device.cu
@@ -21,14 +21,14 @@ void cudaBlurKernel(const float *raw_data, const float *blur_v, float *out_data,
     // as the CPU implementation, as a reference.
     unsigned int thread_index = blockIdx.x * blockDim.x + threadIdx.x;
     unsigned int i;
-    while(thread_index < blur_v_size)
+    while(thread_index < (unsigned int)blur_v_size)
     {
         i = thread_index;
         for (unsigned int j = 0; j <= thread_index; j++)
                 out_data[i] += raw_data[i - j] * blur_v[j];
         thread_index += (blockDim.x * gridDim.x);
     }
-    while(thread_index < n_frames)
+    while(thread_index < (unsigned int)n_frames)
     {
         i = thread_index;
         for (unsigned int j = 0; j < (unsigned int)blur_v_size; j++)
